Added Gperf command list file.
This file allows the automatic creation (using Gperf) of a perfect hash function
that will replace the DFA generated by gendfa.pl.  With larger command sets,
Gperf generates tables that are a bit more compact.  It is also easier to
maintain than the perl script, which required the commands to be ordered just as
in uftps.h.