New chroot emulation mechanism.
Instead of forbidding paths containing "." or ".." as one of its components, a
simple path parser has been added.  This parser eats up the "." and ".."
components, while preserving the parent directories (of the current, i.e.
serving directory) unreachable.

This is an initial version which may need some revision and clean up.