Re-structured source files.
Given the difficulty to group functions in files under an acceptable criterium,
the final decision has been to separate each important function in its own file.

Helper functions will be private and will reside in the same file where they are
used.  If a function is needed in several places, it will become an UFTPS
function and will be separated in its own source file.