New function, open_data_connection.
As the name states, the function handles the opening of data channels before
listing and file transfers.  In the case of active connections (PORT command),
the connection is no longer performed after receiving the PORT request, but when
requrired by a LIST/NLST or RETR request.

Other minor rearrangements have been performed too.