Removed passive port catching.
Recent tests have shown that caching the socket used to accept passive data
connections has some problemas attached.  In particular, it depends a lot on how
clever the client is.

Adding to this the fact that it is possible to let the kernel choose a random
port for you, the PASV command implementation becomes a bit simpler.

Besides, all the other FTP servers do it and does not seem to incur in a
critical performance penalty.