Generic RETR implementation added and other stuff too.
The Makefile has been modified to hold severla send_file (RETR) implementations
to be selected at compile time; as well as being able to compile the Hasefroch
binary too.

The uftps.c file has been renamed to main-unix.c which will be common to all
UNIX versions.  For Hasefroch, there will be a main-hase.c file that will deal
with the specifics of process spawning and socket sharing.