Initial port to Hasefroch.
First port working in hasefroch.  There are still some issues but the basics
seem to work well.  Finally a way to inherit sockets in Hasefroch was found.  As
usual, the solution is copy & paste.

The send_file-*.c sources no longer deal with SS.file_offset all the time, they
copy the value to a local variable for readability and (possibly) for
performance).