Mmap version of send_file ported to Hasefroch.
After learning about Hasefroch File Mappings, it was possible to have a
mmap-like implementation for Hasefroch.  As the code was identical except for
the operating system requests, the new version has been added in place with the
UNIX version.

So now two multiplatform versions are available: generic and mmap.

Hasefroch versions (hase and mmap) still need testing.