Fixed Hasefroch RETR implementation.
Testing revealed that TransmitFile returns FALSE with a WSA_IO_PENDING error
code (like EAGAIN for errno).  It also showed that GetOverlappedResult was not
the function to use, as the wait has to be performed on the socket, not the
file.

An example from the Wine test suite provided the hint to use
WSAGetOverlappedResult for sockets.  Now it works as desired.