Changed distribution and binary generation method.
A new way of generating source packages has been created to help the creation of
binaries for the most common desktop platforms.

It is basically a new makefile target that includes the binary generation script
that was previously executed directly from a repository.  Now is no longer
necessary, but it is not possible to generate binaries from the repository, you
need to invoke the new "bdist" target before.